Server: SQLDatabaseConnection Bug fix: 	* Weird bug of not commiting changes (issue: #225) 	* Cant add catalog product with real barcode (issue: #179) 	* The SQL allow to remove product that still for sale (issue: #180) 	* Moving product packages operations (issue: #182)
Related issue: #31

+++ Server/src/main/java/SQLDatabase/SQLDatabaseEntities.java
@@ -33,7 +33,7 @@ class SQLDatabaseEntities {
 	 */
 	private static final String TYPE_TEXT = "text";
 	private static final String TYPE_INTEGER = "integer";
-	private static final String TYPE_LONG = "integer";
+	private static final String TYPE_LONG = "bigint";
 	private static final String TYPE_REAL = "real";
 	private static final String TYPE_DATE = "datetime";
 	private static final String TYPE_BOOLEAN = "integer";
